role_migration_system/requirements.txt
discord.py>=2.3.0
pandas>=2.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
ratelimit>=2.2.1
openai>=1.0.0
langchain>=0.1.0
jupyter>=1.0.0
ipykernel>=6.0.0